% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_measurement_error.R
\name{add_temp_reference}
\alias{add_temp_reference}
\title{Add temperature from a reference to a hourly_temp object}
\usage{
add_temp_reference(x, ref)
}
\arguments{
\item{x}{hourly_temp object}

\item{ref}{hourly_temp object with the closest reference temperature}
}
\value{
hourly_temp object with the reference temperature in columns temp_r
}
\description{
Add temperature from a reference to a hourly_temp object
}
