% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_measurement_error.R
\name{calc_temp_error}
\alias{calc_temp_error}
\title{Add temperature from a reference to citizen weather stations observations}
\usage{
calc_temp_error(cws, ref)
}
\arguments{
\item{cws}{sf object with citizen weather stations}

\item{ref}{sf object with reference weather stations}
}
\value{
hourly_temp object with the reference temperature in columns temp_r
}
\description{
Add temperature from a reference to citizen weather stations observations
}
