% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_cws.R
\name{calib_cws}
\alias{calib_cws}
\title{Calibrate Personal weather stations with stations from a
reference network}
\usage{
calib_cws(ref, cws, max_dist = 5000)
}
\arguments{
\item{ref}{sf. Reference stations.}

\item{cws}{sf. Personal weather stations.}

\item{max_dist}{integer. Maximum distance allowed between ref
and cws stations.}
}
\description{
Calibrate Personal weather stations with stations from a
reference network
}
\author{
Eva Marques
}
