% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cws.R
\name{clean_cws}
\alias{clean_cws}
\title{Clean Weather Underground data with CrowdQC+. Keep only observations
passing all tests (o3 level)}
\usage{
clean_cws(x, ...)
}
\arguments{
\item{x}{sf with columns: site_id, temp, lat, lon, time}
}
\value{
cleaned data.frame
}
\description{
Clean Weather Underground data with CrowdQC+. Keep only observations
passing all tests (o3 level)
}
\references{
Fenner, D., Bechtel, B., Demuzere, M., Kittner, J. and Meier, F. (2021): CrowdQC+ – A quality-control for crowdsourced air-temperature observations enabling world-wide urban climate applications. Frontiers in Environmental Science 9: 720747. DOI: 10.3389/fenvs.2021.720747.
}
