% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cws.R
\name{clean_cws_large}
\alias{clean_cws_large}
\title{Clean Weather Underground data with CrowdQC+ on an area >100km*100km}
\usage{
clean_cws_large(x, area, epsg_m = "epsg:32119", res = 1e+05, ...)
}
\arguments{
\item{x}{sf of formatted weather station data
(with columns: site_id, temp, lat, lon, time)}

\item{area}{polygon}

\item{epsg_m}{crs in meters (default: epsg:32119)}

\item{res}{resolution of the squares in meters (default: 100km)}
}
\value{
cleaned data.frame
}
\description{
Clean Weather Underground data with CrowdQC+ on an area >100km*100km
}
