% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_wu_inventory.R
\name{create_wu_inventory}
\alias{create_wu_inventory}
\title{Create Weather Underground inventory}
\usage{
create_wu_inventory(fpath, invpath)
}
\arguments{
\item{fpath}{the path to the Weather Underground data files}

\item{invpath}{the path to the output inventory file}
}
\value{
a data.frame with the inventory
}
\description{
Create an inventory of Weather Underground data files with
content metadata (station ID, latitude, longitude, start and end timestamps)
}
