% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cws.R
\name{cut_area}
\alias{cut_area}
\title{Format an area to be used in clean_cws_large (cut in squares of res*res m2)}
\usage{
cut_area(area, epsg_m = "epsg:32119", res = 1e+05)
}
\arguments{
\item{area}{polygon}

\item{epsg_m}{crs in meters (default: epsg:32119)}

\item{res}{resolution of the squares in meters (default: 100km)}
}
\value{
sf of polygons
}
\description{
Format an area to be used in clean_cws_large (cut in squares of res*res m2)
}
