% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghcnh.R
\name{download_ghcnh}
\alias{download_ghcnh}
\title{Download all GHCN-H stations data in area between two dates}
\usage{
download_ghcnh(ts, te, area)
}
\arguments{
\item{ts}{start date}

\item{te}{end date}

\item{area}{a sf, sfc, SpatRaster or SpatVector object}
}
\value{
a data.frame with the GHCN-H stations observations in the area
}
\description{
Download all GHCN-H stations data in area between two dates
}
