% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghcnh.R
\name{download_ghcnh_station}
\alias{download_ghcnh_station}
\title{Download GHCN-H station data of a given year}
\usage{
download_ghcnh_station(site_id, year)
}
\arguments{
\item{site_id}{the GHCN-H station ID}

\item{year}{the year of the data}
}
\value{
a data.frame with the GHCN-H station raw data
}
\description{
Download GHCN-H station data of a given year
}
\author{
Eva Marques
}
