% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_purpleair.R
\name{download_pa}
\alias{download_pa}
\title{Download all PurpleAir stations data in area between two dates}
\usage{
download_pa(ts, te, area, api_key)
}
\arguments{
\item{ts}{start date}

\item{te}{end date}

\item{area}{a sf, sfc, SpatRaster or SpatVector object}
}
\description{
Download all PurpleAir stations data in area between two dates
}
