% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_measurement_error.R
\name{est_temp_error}
\alias{est_temp_error}
\title{Add temperature from a reference to citizen weather stations observations}
\usage{
est_temp_error(cws, ref)
}
\arguments{
\item{cws}{sf (or inherited) object with citizen weather stations
with columns site_id and geometry}

\item{ref}{sf (or inherited) object with reference weather stations
with columns site_id and geometry}
}
\value{
sf (or inherited) object with additional columns temp_ref, temp_err,
ref_id, dist_to_ref
}
\description{
Add temperature from a reference to citizen weather stations observations
}
\author{
Eva Marques
}
