% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_measurement_error.R
\name{estimate_temp_error}
\alias{estimate_temp_error}
\title{Estimate measurement errors for a hourly_temp object}
\usage{
estimate_temp_error(x, ref)
}
\description{
Estimate measurement errors for a hourly_temp object
}
