% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_measurement_error.R
\name{find_closest_ref}
\alias{find_closest_ref}
\title{Find the closest reference site for each site in a cws object}
\usage{
find_closest_ref(cws, ref)
}
\arguments{
\item{cws}{sf (or inherited) object with citizen weather stations
with columns site_id and geometry}

\item{ref}{sf (or inherited) object with reference weather stations
with columns site_id and geometry}
}
\value{
sf object with the closest reference site for each cws
with additional columns ref_id and dist_to_ref
}
\description{
Find the closest reference site for each site in a cws object
}
