% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghcnh.R
\name{find_ghcnh_polygon}
\alias{find_ghcnh_polygon}
\title{Find all the GHCN-H stations within a polygon}
\usage{
find_ghcnh_polygon(polygon)
}
\arguments{
\item{polygon}{a sf object with a POLYGON geometry}
}
\value{
a sf object with the GHCN-H stations within the polygon
}
\description{
Find all the GHCN-H stations within a polygon
}
\author{
Eva Marques
}
