% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnh.R
\name{find_nearest_ghcnh}
\alias{find_nearest_ghcnh}
\title{Find the nearest GHCN-H station to a point}
\usage{
find_nearest_ghcnh(lat, lon)
}
\arguments{
\item{lat}{latitude of the point (in WGS84)}

\item{lon}{longitude of the point (in WGS84)}
}
\value{
a sf object with the nearest GHCN-H station
}
\description{
Find the nearest GHCN-H station to a point
}
\author{
Eva Marques
}
