% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_purpleair.R
\name{find_sensors}
\alias{find_sensors}
\title{Find a list of sensors in a specific area}
\usage{
find_sensors(nwlat, selat, nwlng, selng, api_key, location_type = 0)
}
\arguments{
\item{nwlat}{North-west latitude}

\item{selat}{South-east latitude}

\item{nwlng}{North-west longitude}

\item{selng}{South-east longitude}

\item{api_key}{API key for PurpleAir}

\item{location_type}{Location type (0 = outside, 1 = inside)}
}
\value{
A data frame with sensors information
}
\description{
Find a list of sensors in a specific area
}
\references{
Code adapted from Callahan J, Martin H, Wilson K, Brasel T, Miller H (2023). AirSensor: Process and Display Data from Air Quality Sensors. R package version 1.1.1, https://CRAN.R-project.org/package=AirSensor.
}
