% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_observations.R
\name{format_ghcnh}
\alias{format_ghcnh}
\title{Format observations from raw data downloaded on NOAA website
See function load_ghcnh_station.}
\usage{
format_ghcnh(raw)
}
\arguments{
\item{raw}{a data.frame, data.table, sf or sftime with the raw observations
and columns "Year", "Month", "Day", "Hour", "temperature", "Latitude",
"Longitude", "temperature_Source_Code"}
}
\value{
sftime from hourly_temp class
}
\description{
Format observations from raw data downloaded on NOAA website
See function load_ghcnh_station.
}
\author{
Eva Marques
}
