% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_observations.R
\name{format_pa}
\alias{format_pa}
\title{Format observations directly downloaded on PurpleAir API.}
\usage{
format_pa(raw, raw_tz = "UTC", raw_temp_unit = "F", raw_crs = 4326)
}
\arguments{
\item{raw}{a data.frame, data.table, sf with the raw observations
and columns "time_stamp", "latitude", "longitude", "temperature"}

\item{raw_tz}{the initial timezone, see PurpleAir API documentation}

\item{raw_temp_unit}{the initial temperature unit}

\item{raw_crs}{the initial coordinate reference system}
}
\value{
sf from hourly_temp class
}
\description{
Format observations directly downloaded on PurpleAir API.
}
\author{
Eva Marques
}
