% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_observations.R
\name{format_wu}
\alias{format_wu}
\title{Format observations sent by IBM.}
\usage{
format_wu(raw, raw_tz = "UTC", raw_temp_unit = "F", raw_crs = 4326)
}
\arguments{
\item{raw}{a data.frame, data.table, sf or sftime with the raw observations
and columns "obsTimeUtc", "lat", "lon", "tempAvg"}

\item{raw_tz}{the initial timezone, see PurpleAir API documentation}

\item{raw_temp_unit}{the initial temperature unit}

\item{raw_crs}{the initial coordinate reference system}
}
\value{
sftime from hourly_temp class
}
\description{
Format observations sent by IBM.
}
\author{
Eva Marques
}
