% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_unique_site_id.R
\name{generate_site_id}
\alias{generate_site_id}
\title{Generate unique sensor ID}
\usage{
generate_site_id(x)
}
\arguments{
\item{x}{A data.frame, data.table, sf or sftime with columns "lat" and "lon"}
}
\value{
A data frame with an additional column "id"
}
\description{
This function generates a unique sensor "id" column based on
the latitude and longitude of the sensor.
}
\author{
Eva Marques
}
