% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plots.R
\name{hourly_boxplot}
\alias{hourly_boxplot}
\title{Hourly boxplot of temperature for per network (network)}
\usage{
hourly_boxplot(data, ts, te, var)
}
\arguments{
\item{data}{data frame or sftime with all the network, contains "network"
column}

\item{ts}{datetime, start time}

\item{te}{datetime, end time}

\item{var}{continuous variable to boxplot (in general: temp)}
}
\value{
Hourly boxplots per network
}
\description{
Hourly boxplot of temperature for per network (network)
}
\author{
Eva Marques
}
