% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cws.R
\name{hourly_boxplot_networks}
\alias{hourly_boxplot_networks}
\title{Hourly boxplot of temperature for per network (source)}
\usage{
hourly_boxplot_networks(data, var)
}
\arguments{
\item{data}{data frame or sftime with all the network, contains "source"
column}

\item{var}{continuous variable to boxplot (in general: temp)}
}
\value{
A ggplo2 object with the boxplots
}
\description{
Hourly boxplot of temperature for per network (source)
}
\author{
Eva Marques
}
