% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_hourly_temp.R
\name{hourly_temp}
\alias{hourly_temp}
\title{Build a hourly_temp from a data.frame, data.table, sf or sftime}
\usage{
hourly_temp(x, temp = "temp", lat = "lat", lon = "lon", time = "time", network)
}
\arguments{
\item{x}{a data.frame, data.table, sf or sftime}

\item{temp}{the column name for the temperature in degree celcius}

\item{lat}{the column name for the latitude}

\item{lon}{the column name for the longitude}

\item{time}{the column name for the time in UTC}

\item{network}{the name of the network}
}
\value{
a hourly_temp object
}
\description{
Build a hourly_temp from a data.frame, data.table, sf or sftime
}
\author{
Eva Marques
}
