% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnh.R
\name{load_ghcnh_station}
\alias{load_ghcnh_station}
\title{Load GHCN-H station data of a given year}
\usage{
load_ghcnh_station(site_id, year)
}
\arguments{
\item{site_id}{the GHCN-H station ID}

\item{year}{the year of the data}
}
\value{
a data.frame with the GHCN-H station raw data
}
\description{
Load GHCN-H station data of a given year
}
\author{
Eva Marques
}
