% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_purpleair.R
\name{load_pa}
\alias{load_pa}
\title{If a file is provided, open data from file. If not, call download_pa().}
\usage{
load_pa(ts, te, area, storage_file = NULL, api_key = NULL)
}
\arguments{
\item{ts}{start date}

\item{te}{end date}

\item{area}{a sf, sfc, SpatRaster or SpatVector object}

\item{storage_file}{file path where PurpleAir data is stored}

\item{api_key}{API key for PurpleAir}
}
\description{
If a file is provided, open data from file. If not, call download_pa().
}
