% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_weatherunderground.R
\name{load_wu}
\alias{load_wu}
\title{Load Weather Underground stations for a given period and area.}
\usage{
load_wu(ts, te, area, inventory)
}
\arguments{
\item{ts}{Start time}

\item{te}{End time}

\item{area}{Area of interest}

\item{inventory}{an sf object with inventory of}
}
\value{
a data.table with the raw WU data
}
\description{
Load Weather Underground stations for a given period and area.
}
\author{
Eva Marques
Weather Underground stations (see create_wu_inventory function)
}
