% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_weatherunderground.R
\name{load_wu_period}
\alias{load_wu_period}
\title{Load Weather Underground stations for a given period.}
\usage{
load_wu_period(rpath, ts, te, wpath)
}
\arguments{
\item{rpath}{Path to the directory containing the Weather Underground
station files of a region}

\item{ts}{Start time}

\item{te}{End time}
}
\value{
A data.table containing the weather data for the given period
}
\description{
Load Weather Underground stations for a given period.
}
\author{
Eva Marques
}
