% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cws.R
\name{manage_na}
\alias{manage_na}
\title{Remove NA and stations with too many NA}
\usage{
manage_na(data, na_thresh = 0.1)
}
\arguments{
\item{data}{formatted sf (with columns: site_id, temp, time)}

\item{na_thresh}{threshold of NA to remove a station (0 <= na_tresh <= 1)}
}
\value{
cleaned data.frame
}
\description{
Remove NA and stations with too many NA
}
