% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plots.R
\name{map_stations}
\alias{map_stations}
\title{map observations with imperviousness}
\usage{
map_stations(
  data,
  datetime,
  var,
  imp,
  palname = "temp_ipcc",
  netw_shape = c(WU = 17, PA = 18, `NCEI/ASOS/AWOS` = 16, `NCEI/US CRN` = 15),
  title
)
}
\arguments{
\item{data}{data frame or sftime of cws observations}

\item{datetime}{datetime to plot}

\item{var}{continuous variable to plot (in general: temp)}

\item{imp}{imperviousness raster}

\item{palname}{name of the palette to use}

\item{netw_shape}{shape values for the network}
}
\description{
map observations with imperviousness
}
