% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_pal}
\alias{my_pal}
\title{Find my home-made palettes}
\usage{
my_pal(pal_name)
}
\arguments{
\item{pal_name}{Name of the palette}
}
\value{
A vector with colors
}
\description{
Find my home-made palettes
}
\author{
Eva Marques
}
