% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cws.R
\name{plot_ts}
\alias{plot_ts}
\title{Timeseries plot of temperature}
\usage{
plot_ts(data, ts, te)
}
\arguments{
\item{data}{data frame or sftime of cws observations}

\item{ts}{datetime, start time}

\item{te}{datetime, end time}
}
\value{
A ggplo2 object with the timeseries
}
\description{
Timeseries plot of temperature
}
\author{
Eva Marques
}
