% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cws.R
\name{plot_ts_ref}
\alias{plot_ts_ref}
\title{Timeseries plot of temperature with reference network}
\usage{
plot_ts_ref(data, ref, ts, te)
}
\arguments{
\item{data}{data frame or sftime of cws observations}

\item{ref}{data frame or sftime, reference network}

\item{ts}{datetime, start time}

\item{te}{datetime, end time}
}
\value{
A ggplo2 object with the timeseries
}
\description{
Timeseries plot of temperature with reference network
}
\author{
Eva Marques
}
