% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_observations.R
\name{remove_daily_cws}
\alias{remove_daily_cws}
\title{Remove personal weather stations which provide daily statistics only.}
\usage{
remove_daily_cws(x)
}
\arguments{
\item{x}{data.frame or sf. Personal weather stations dataset.}
}
\value{
same as x without daily stations.
}
\description{
Remove personal weather stations which provide daily statistics only.
}
\author{
Eva Marques
}
