% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cws.R
\name{rename_for_plot}
\alias{rename_for_plot}
\title{Rename x columns for homogeneous plots}
\usage{
rename_for_plot(x, time, temp, id, lat, lon)
}
\arguments{
\item{x}{data frame or sftime}

\item{time}{character, name of the time column}

\item{temp}{character, name of the temperature column}

\item{id}{character, name of the id column}

\item{lat}{character, name of the latitude column}

\item{lon}{character, name of the longitude column}
}
\value{
same as x with renamed columns
}
\description{
Rename x columns for homogeneous plots
}
\author{
Eva Marques
}
