% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_purpleair.R
\name{request_sensors_history}
\alias{request_sensors_history}
\title{Load PurpleAir sensors history data}
\usage{
request_sensors_history(
  nwlat,
  selat,
  nwlng,
  selng,
  location_type = 0,
  start_ts,
  end_ts,
  api_key,
  average = "60",
  fields = "temperature, humidity"
)
}
\arguments{
\item{nwlat}{North-west latitude}

\item{selat}{South-east latitude}

\item{nwlng}{North-west longitude}

\item{selng}{South-east longitude}

\item{location_type}{Location type}

\item{start_ts}{Start timestamp}

\item{end_ts}{End timestamp}

\item{api_key}{API key for PurpleAir}

\item{average}{Average time in minutes}

\item{fields}{Fields to be included in the data}
}
\value{
A data frame with sensors history data
}
\description{
Load PurpleAir sensors history data
}
\references{
Code adapted from Callahan J, Martin H, Wilson K, Brasel T, Miller H (2023). AirSensor: Process and Display Data from Air Quality Sensors. R package version 1.1.1, https://CRAN.R-project.org/package=AirSensor.
}
