% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cws.R
\name{tile_ts}
\alias{tile_ts}
\title{Tile plot with temperature through time for all cws}
\usage{
tile_ts(data)
}
\arguments{
\item{data}{data frame or sftime of cws observations}
}
\value{
A ggplo2 object with the tileplot
}
\description{
Tile plot with temperature through time for all cws
}
\author{
Eva Marques
}
