% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plots.R
\name{timeseries}
\alias{timeseries}
\title{Timeseries plot of temperature}
\usage{
timeseries(data, ts, te, var)
}
\arguments{
\item{data}{data frame or sftime of cws observations}

\item{ts}{datetime, start time}

\item{te}{datetime, end time}

\item{var}{continuous variable to plot (in general: temp)}
}
\value{
Timeseries plot
}
\description{
Timeseries plot of temperature
}
\author{
Eva Marques
}
