% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_cws.R
\name{calib_cws}
\alias{calib_cws}
\title{Calibration of citizen weather station with regard to a reference network}
\usage{
calib_cws(x, ref, max_dist = 10000)
}
\arguments{
\item{x}{a data.frame with the observations to calibrate}

\item{ref}{a data.frame with the reference observations}

\item{max_dist}{the maximum distance in meters to consider
a station as a reference (default = 10000)}
}
\value{
a list with the calibrated observations, the average bias,
the number of reference stations used to compute the bias, and the
start and end time of the observations. Improvement idea: time constraint,
bias calculation per month / season / on a sliding temporal window
}
\description{
Calibration of citizen weather station with regard to a reference network
}
