% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_cws.R
\name{calib_cws}
\alias{calib_cws}
\title{Calibration of citizen weather station with regard to a reference network}
\usage{
calib_cws(x, ref, max_dist = 10000)
}
\arguments{
\item{x}{sf (or inherited) object with citizen weather stations to calibrate
with columns site_id and geometry}

\item{ref}{sf (or inherited) object with reference weather stations
with columns site_id and geometry}

\item{max_dist}{the maximum distance in meters to consider
a station as a reference (default = 10000)}
}
\value{
a list with the calibrated observations, the average bias,
the number of reference stations used to compute the bias, and the
start and end time of the observations.
}
\description{
Calibration of citizen weather station with regard to a reference network
}
