% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cws.R
\name{clean_cws}
\alias{clean_cws}
\title{Clean Weather Underground data with CrowdQC+}
\usage{
clean_cws(x)
}
\arguments{
\item{x}{data.frame with columns: site_id, temp, lat, lon, time}
}
\value{
cleaned data.frame
}
\description{
Clean Weather Underground data with CrowdQC+
}
