% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_purpleair_from_api.R
\name{request_sensor_history}
\alias{request_sensor_history}
\title{Load PurpleAir sensor history data}
\usage{
request_sensor_history(
  start_ts,
  end_ts,
  sensor_index,
  api_key,
  average = "60",
  fields = "temperature, humidity"
)
}
\arguments{
\item{start_ts}{Start timestamp}

\item{end_ts}{End timestamp}

\item{sensor_index}{Sensor index}

\item{api_key}{API key for PurpleAir}

\item{average}{Average time in minutes}

\item{fields}{Fields to be included in the data}
}
\value{
A data frame with sensor history data
}
\description{
Load PurpleAir sensor history data
}
