% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_observations.R
\name{summarize_hourly_temp}
\alias{summarize_hourly_temp}
\title{Summarize hourly temperature at each site.}
\usage{
summarize_hourly_temp(x, time, temp, lat, lon)
}
\arguments{
\item{x}{a data.frame, data.table, sf or sftime}

\item{time}{the column name for the time}

\item{temp}{the column name for the temperature}

\item{lat}{the column name for the latitude}

\item{lon}{the column name for the longitude}
}
\value{
a data.frame of hourly average of temperature at each site
}
\description{
Summarize hourly temperature at each site.
}
\author{
Eva Marques
}
