% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plots.R
\name{tile_ts}
\alias{tile_ts}
\title{Tile plot with temperature through time for all cws}
\usage{
tile_ts(data, ts, te, var, palname = "temp_ipcc")
}
\arguments{
\item{data}{data frame or sftime of cws observations}

\item{ts}{datetime, start time}

\item{te}{datetime, end time}

\item{var}{continuous variable to plot (in general: temp)}

\item{palname}{name of the palette to use}
}
\value{
Tile plot
}
\description{
Tile plot with temperature through time for all cws
}
\author{
Eva Marques
}
